Suppresse dead_code warnings about struct fields
Mark lua_Debug with #[repr(C)], even though that hasn't been implemented
yet, as it suppresses the dead_code warning about the Lua internal field
in the struct.

Rename State.stackspace to _stackspace to suppress the warning there.
That field is never read directly, the struct is always transmuted to
ExternState or RawState first, so it was considered dead.